+++ src/pyzet/exceptions.py
@@ -1,0 +1,6 @@
+"""Custom exceptions mainly used for very specific control flow."""
+from __future__ import annotations
+
+
+class CreateNewZettel(Exception):
+    pass