+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+* Ask to create a new zettel if none are found by grep in `pyzet edit` & `pyzet show`.
+
 ### Changed
 
 * Zero-padding is used when displaying the list of matched zettels.